Split base script into chunks for easier management. Removed unused constructor parameter
