+++ source/LiteMigrator/LiteMigration.cs
@@ -3,17 +3,13 @@
  * Author:  Damian Suess
  * File:    LiteMigrator.cs
  * Description:
- *  Simple SQL Migration Engine
+ *  Simple SQL Migration Engine base class and constructors
  */
 
 using System;
-using System.Collections.Generic;
 using System.Reflection;
-using System.Threading.Tasks;
-using LiteMigrator.DataObjects;
 using LiteMigrator.Engines;
 using LiteMigrator.Factory;
-using LiteMigrator.Versioning;
 using SQLite;
 
 namespace LiteMigrator;
@@ -32,41 +28,45 @@ public partial class LiteMigration : IDisposable
   private string _databasePath = string.Empty;
   private bool _disposed = false;
   private bool _isInitialized = false;
+
+  /* TODO: Allow user to choose different engines (i.e. SQLite or SQLCypher)
   private IEngine _sqlEngine;
+  */
 
   /// <summary>
-  ///   Initializes a new instance of the <see cref="LiteMigration"/> class.
-  ///   Assumes in-memory database, current namespace, and using generic SQLite.</summary>
-  /// <param name="databaseType">Type of database.</param>
+  ///   Initializes a new instance of the <see cref="LiteMigration"/> class using an in-memory database.
+  ///   Assumes the current namespace, and using generic SQLite.
+  /// </summary>
   public LiteMigration()
-    : this(InMemoryDatabase, string.Empty, DatabaseType.SQLite)
+    : this(InMemoryDatabase, string.Empty) ////, DatabaseType.SQLite)
   {
     // Set to current namespace, it's a something
     BaseNamespace = GetType().Namespace;
   }
 
   /// <summary>
-  ///   Initializes a new instance of the <see cref="LiteMigration"/> class.
+  ///   Initializes a new instance of the <see cref="LiteMigration"/> class using an in-memory database.
   ///   Assumes in-memory database and using generic SQLite.
   /// </summary>
   /// <param name="baseNamespace">Namespace to scripts.</param>
   public LiteMigration(string baseNamespace)
-    : this(InMemoryDatabase, baseNamespace, DatabaseType.SQLite)
+    : this(InMemoryDatabase, baseNamespace) ////, DatabaseType.SQLite)
   {
   }
 
   /// <summary>
-  ///   Initializes a new instance of the <see cref="LiteMigration"/> class.
+  ///   Initializes a new instance of the <see cref="LiteMigration"/> class using an in-memory database.
   ///   Assumes in-memory database and the current namespace contains the scripts.
   /// </summary>
   /// <param name="assm">Resource file with migration scripts.</param>
   /// <param name="baseNamespace">Assembly path to migration scripts.</param>
   public LiteMigration(Assembly assm, string baseNamespace)
-    : this(InMemoryDatabase, baseNamespace, DatabaseType.SQLite)
+    : this(InMemoryDatabase, baseNamespace) ////, DatabaseType.SQLite)
   {
     Migrations.BaseAssemblyFile = assm.Location;
   }
 
+  /*
   /// <summary>
   ///   Initializes a new instance of the <see cref="LiteMigration"/> class.
   ///   Assumes in-memory database and the current namespace contains the scripts.
@@ -78,6 +78,7 @@ public partial class LiteMigration : IDisposable
     // Set to current namespace, it's a something
     BaseNamespace = GetType().Namespace;
   }
+  */
 
   /// <summary>
   ///   Initializes a new instance of the <see cref="LiteMigration"/> class.
@@ -87,7 +88,7 @@ public partial class LiteMigration : IDisposable
   /// <param name="assm">Resource file with migration scripts.</param>
   /// <param name="baseNamespace">Namespace to scripts.</param>
   public LiteMigration(string databasePath, Assembly assm, string baseNamespace)
-    : this(databasePath, baseNamespace, DatabaseType.SQLite)
+    : this(databasePath, baseNamespace) ////, DatabaseType.SQLite)
   {
     Migrations.BaseAssemblyFile = assm.Location;
   }
@@ -99,12 +100,13 @@ public partial class LiteMigration : IDisposable
   /// <param name="baseNamespace">Namespace to scripts.</param>
   /// <param name="databaseType">Type of database connection.</param>
   /// <param name="baseAssembly">Migration's base assembly name.</param>
-  public LiteMigration(string databasePath, string baseNamespace, DatabaseType databaseType, string baseAssembly = "")
+  //// public LiteMigration(string databasePath, string baseNamespace, DatabaseType databaseType, string baseAssembly = "")
+  public LiteMigration(string databasePath, string baseNamespace, string baseAssembly = "")
   {
     ////RevisionTable = nameof(VersionInfo);  // FUTURE
     // Set to current namespace, it's a something
     DatabasePath = databasePath;
-    DatabaseType = databaseType;
+    //// DatabaseType = databaseType;
 
     // Create version info table here
     // Initialize().Wait();
@@ -114,6 +116,7 @@ public partial class LiteMigration : IDisposable
       BaseNamespace = baseNamespace,
     };
 
+    /*
     switch (databaseType)
     {
       case DatabaseType.SQLiteCipher:
@@ -128,6 +131,7 @@ public partial class LiteMigration : IDisposable
         _sqlEngine = new Engines.SqlitePclEngine();
         break;
     }
+    */
 
     // Connect to database immediately
     if (!Connect(DatabasePath))
@@ -189,7 +193,7 @@ public partial class LiteMigration : IDisposable
   {
     get
     {
-      if (_connection is null)
+      if (Connection is null)
         return false;
       else
         return true;
@@ -240,7 +244,7 @@ public partial class LiteMigration : IDisposable
       return;
 
     // Close database connection.
-    _connection?.CloseAsync();
+    Connection?.CloseAsync();
   }
 
   //// FUTURE: Consider adding this feature one day
@@ -248,350 +252,4 @@ public partial class LiteMigration : IDisposable
   //// public string RevisionTable { get; set; }
 
   ////public VersionFactory Versions { get; private set; }
-}
-
-/// <summary>LiteMigration, migration scripts.</summary>
-public partial class LiteMigration
-{
-  /*
-  /// <summary>Executes a single migration script.</summary>
-  /// <param name="migration">Migration script information.</param>
-  /// <returns>True if executed without errors.</returns>
-  [Obsolete("Not implemented.")]
-  public async Task<bool> MigrateUpAsync(IMigration migration)
-  {
-    await Task.Yield();
-    throw new NotImplementedException();
-  }
-  */
-
-  /// <summary>
-  ///   Executes all missing migration scripts inside a database transaction
-  ///   so it can rollback all upgrades upon failure.
-  ///
-  ///   OnSuccess: Remove backup / commit changes.
-  ///   OnFailure: Rollback upon failure.
-  /// </summary>
-  /// <returns>True if executed without errors.</returns>
-  public async Task<bool> MigrateUpAsync()
-  {
-    bool hasError = false;
-    string errMsg = string.Empty;
-
-    bool initOk = await ReinitializeAsync();
-
-    if (!initOk)
-    {
-      hasError = true;
-      errMsg = "Could not initialize migration VersionInfo table";
-    }
-    else
-    {
-      var allMigs = Migrations.GetSortedMigrations();
-      var missing = await GetMissingMigrationsAsync();
-
-      // _sqlEngine.Connect(DatabasePath);
-      //// SQLiteAsyncConnection db = new SQLiteAsyncConnection(DatabasePath);
-
-      foreach (var mig in missing)
-      {
-        string path = mig.Value.Script;
-        string migScript = Migrations.GetMigrationScriptByResource(path);
-        var parser = new ParserFactory(migScript);
-
-        // isOk = _sqlEngine.ExecuteMigration(mig, out errMsg);
-        // if (!isOk) break;
-        await _connection.RunInTransactionAsync(tran =>
-        {
-          // Cannot begin a transaction while in one
-          // tran.BeginTransaction();
-          try
-          {
-            //// Only executes the 1st statement
-            //// int x = tran.Execute(query);
-
-            int cmdCount = 0;
-            string lastQuery = string.Empty;
-            string currentQuery = string.Empty;
-
-            try
-            {
-              // Split commands by ';'
-              var commands = parser.GetCommands(migScript);
-
-              // Belongs in ParseFactory
-              int beginCount = 0;
-              string cmdCache = string.Empty;
-
-              // Execute each command
-              // TODO: Move into ParseFactory
-              foreach (var query in commands)
-              {
-                if (!parser.IsCommand(query))
-                  continue;
-
-                parser.Concat(query);   // cmdCache += query;
-
-                if (parser.CurrentCommand.Contains("CREATE TRIGGER"))
-                {
-                  // Make sure every BEGIN has an END before continuing !
-                  if (query.Contains("BEGIN"))
-                    beginCount++;
-
-                  if (query.Contains("END;"))
-                    beginCount--;
-
-                  if (beginCount > 0)
-                    continue;
-                }
-
-                ++cmdCount;
-
-                // Increment the execution line
-                parser.CountLines(query);
-                currentQuery = parser.CurrentCommand;   // currentQuery = cmdCache;
-
-                var cmd = tran.CreateCommand(currentQuery);
-                cmd.ExecuteNonQuery();
-                parser.ClearCommand();
-
-                lastQuery = currentQuery;
-              }
-            }
-            catch (Exception execEx)
-            {
-              string err = $"Migration failed on Command number: '{cmdCount}'{Environment.NewLine}" +
-                           $"* Exception: {execEx.Message}{Environment.NewLine}" +
-                           $"* Line: {parser.Lines}{Environment.NewLine}" +
-                           $"* Path: '{path}'{Environment.NewLine}" +
-                           $"* Query: {currentQuery}{Environment.NewLine}" +
-                           $"* Previous Query: {lastQuery}";
-
-              System.Diagnostics.Debug.WriteLine($"[ERROR] [LiteMigration] [{err}]");
-              LastException = execEx;
-              throw new Exception(err);
-            }
-
-            var verInfo = new VersionInfo()
-            {
-              AppliedDttm = DateTime.Now,
-              Description = mig.Value.Description,
-              VersionNumber = mig.Value.VersionNumber,
-            };
-
-            tran.Insert(verInfo);
-
-            AddVersion(mig.Value);
-
-            tran.Commit();
-          }
-          catch (Exception ex)
-          {
-            hasError = true;
-            errMsg = ex.Message;
-            tran.Rollback();
-
-            // TODO: Report which migration script & error message
-            LastError = errMsg;
-            LastException = ex;
-          }
-        });
-
-        if (hasError)
-          break;
-      }
-
-      // _sqlEngine.Close()
-      //// await _connection.CloseAsync();
-    }
-
-    if (hasError)
-    {
-      // Why did we get an error
-      System.Diagnostics.Debug.WriteLine("LiteMigrator - ERROR - " + LastError);
-    }
-
-    return !hasError;
-  }
-
-  public async Task<bool> ReinitializeAsync()
-  {
-    // Migrations = new MigrationFactory();
-
-    // Maybe use Lazy loading?
-    ////Versions = new VersionFactory();
-    ////Versions.Initialize();
-
-    await VersionInitializeAsync();
-
-    return true;
-  }
-}
-
-/// <summary>Version Factory of LiteMigration.</summary>
-public partial class LiteMigration
-{
-  /// <summary>Gets or sets the list of applied migrations.</summary>
-  /// <value>The list of applied migrations.</value>
-  public Versions Versions { get; set; }
-
-  /////// <summary>
-  /////// Bulk register scripts into VersionInfo table
-  /////// </summary>
-  /////// <param name="migration">Migration script.</param>
-  /////// <returns>Task.</returns>
-  ////public async Task AddVersionsAsync(List<Migration> migrations)
-  ////{
-  ////
-  ////}
-
-  public void AddVersion(long version)
-  {
-    if (Versions.IsApplied(version))
-      return;
-
-    Versions.AddItem(version);
-  }
-
-  public void AddVersion(IMigration migration)
-  {
-    if (Versions.IsApplied(migration.VersionNumber))
-    {
-      // TODO: throw new VersionPreviouslyExistsException();
-      return;
-    }
-
-    // Don't think now is the right time to add the script
-    ////migration.AppliedDttm = System.DateTime.Now;
-    ////SQLiteAsyncConnection db = new SQLiteAsyncConnection(DatabasePath);
-    ////await db.InsertAsync(migration);
-    ////await db.CloseAsync();
-
-    Versions.AddItem(migration.VersionNumber);
-  }
-
-  /// <summary>
-  ///   Read from database the list of installed migration scripts.
-  /// </summary>
-  /// <returns>Sorted list of migration scripts.</returns>
-  public async Task<SortedDictionary<long, IVersionInfo>> GetInstalledMigrationsAsync()
-  {
-    var sorted = new SortedDictionary<long, IVersionInfo>();
-
-    try
-    {
-      //// SQLiteAsyncConnection db = new SQLiteAsyncConnection(DatabasePath);
-      var list = await _connection.Table<VersionInfo>().ToListAsync();
-
-      // TODO: In-memory DB will erase when closed
-      //// await _connection.CloseAsync();
-
-      foreach (VersionInfo item in list)
-      {
-        sorted.Add(item.VersionNumber, item);
-
-        Versions.AddItem(item.VersionNumber);
-      }
-    }
-    catch (SQLiteException ex)
-    {
-      // Causes of failure:
-      //  1) VersionInfo table doesn't exist. This can happen when using the ":memory:" database
-      //  2) Database is locked - command already in progress
-      System.Diagnostics.Debug.WriteLine("[Error] [GetInstalledMigrationsAsync] " + ex.Message);
-
-      LastError = ex.Message;
-
-      // Consider: Just in-case, lets make sure we have a DB
-      // await VersionInitializeAsync();
-    }
-
-    return sorted;
-  }
-
-  /// <summary>Get list of migrations which have not been applied to our database yet.</summary>
-  /// <returns>Sorted dictionary of non-applied migrations.</returns>
-  public async Task<SortedDictionary<long, IMigration>> GetMissingMigrationsAsync()
-  {
-    var deltaList = new SortedDictionary<long, IMigration>();
-    var installed = await GetInstalledMigrationsAsync();
-    var available = Migrations.GetSortedMigrations();
-
-    foreach (var item in available)
-    {
-      if (!installed.ContainsKey(item.Key))
-        deltaList.Add(item.Key, item.Value);
-    }
-
-    return deltaList;
-  }
-
-  /// <summary>Registers migration script.</summary>
-  /// <param name="db">Database connection.</param>
-  /// <param name="migration">Migration information.</param>
-  /// <returns>Task.</returns>
-  public async Task RegisterVersionAsync(SQLiteAsyncConnection db, IMigration migration)
-  {
-    migration.AppliedDttm = System.DateTime.Now;
-
-    var verInfo = new VersionInfo()
-    {
-      AppliedDttm = DateTime.Now,
-      Description = migration.Description,
-      VersionNumber = migration.VersionNumber,
-    };
-
-    await db.InsertAsync(verInfo);
-
-    // Cache up what was saved
-    AddVersion(migration);
-  }
-
-  public CreateTableResult? VersionInitialize()
-  {
-    try
-    {
-      Versions = new Versions();
-      var task = _connection.CreateTableAsync<VersionInfo>();
-
-      // Make sure the table is created before moving along
-      task.Wait();
-      var result = task.Result;
-
-      return result;
-
-      /*
-      SQLiteAsyncConnection db = new SQLiteAsyncConnection(DatabasePath);
-      db.CreateTableAsync<VersionInfo>();
-      db.CloseAsync();
-      */
-    }
-    catch (Exception ex)
-    {
-      LastException = ex;
-      System.Diagnostics.Debug.WriteLine("[Error] [VersionInitialize] " + ex.Message);
-
-      return null;
-    }
-  }
-
-  public async Task VersionInitializeAsync()
-  {
-    try
-    {
-      Versions = new Versions();
-      await _connection.CreateTableAsync<VersionInfo>();
-
-      /*
-      SQLiteAsyncConnection db = new SQLiteAsyncConnection(DatabasePath);
-      await db.CreateTableAsync<VersionInfo>();
-      await db.CloseAsync();
-      */
-    }
-    catch (Exception ex)
-    {
-      LastException = ex;
-      System.Diagnostics.Debug.WriteLine("[Error] [VersionInitialize] " + ex.Message);
-    }
-  }
 }
